Speed up `dispatcher.Handle`
This moves the `Start()` invocation into the go routine so that these can happen in parallel, which dramatically speeds up the GCS dispatch path because all of the Starts can happen concurrently instead of serially (with the slow latency of a GCS commit).

Now the GCS implementation can initiate O(concurrency) tasks virtually instantaneously.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>